cleaner package structure
